crypto_trading_system/requirements.txt
ccxt
pandas
numpy
ta
openai
python-dotenv
matplotlib
requests
yfinance
scikit-learn
torch
chromadb
lightgbm
backtrader
